Update sbom-tool-cli-reference.md
Added a note to the sbom-tool validate command to warn users not to output their validation file in the same folder as the drop path.  The output path needs to be outside of your drop path.  Otherwise on subsequent validations you'll receive a FAILURE even though nothing changed in your project dependencies.  This is because the validate file gets added to your drop path and it's messed up the hash comparison because you added a file into drop path.  Hopefully this makes sense.