fix some incorrect as usages in GetDocumentNamespaceUri
the assumption seems to be that these values must be non nullable strings. since if they are null or not strings, it would result in a null ref inside EscapeDataString.

so it is better to do a cast. and if the assumption is ever incorrect, the exception will occur earlier and be more accurate